Add mapping for Tomcat ServiceRefFactory (WSDL)
This is a partial work whereby which an attempt is made to load a WSDL from an HTTP server and then use this WSDL to trigger RCE. The class org.apache.naming.factory.webservices.ServiceRefFactory has a getObjectInstance method that is capable of loading a WSDL file. Further there exists exploits that can use WSDL in combination with ELProcessor to trigger RCE. However, this is only partial work for now.